+++ src/commands/download.rs
@@ -40,7 +40,7 @@ pub async fn download(args: &DownloadArgs) -> Result<(), RustiqueError> {
     };
 
     // verify game version
-    let game_versions = sorted_game_versions();
+    let game_versions = sorted_game_versions().await;
 
     let user_version = args.game_version.replace('v', "");
     let mut found = false;