+++ src/commands/info.rs
@@ -1,6 +1,7 @@
 use comfy_table::modifiers::UTF8_ROUND_CORNERS;
 use comfy_table::presets::UTF8_FULL_CONDENSED;
 use comfy_table::{CellAlignment, Color, ContentArrangement, Row, Table};
+use crate::api;
 use crate::api::api_structs::Releases;
 use crate::api::client::ApiClient;
 use crate::commands::arg_structs::info_args::ModInfoArgs;
@@ -132,7 +133,7 @@ pub async fn info(args: &ModInfoArgs) -> Result<(), RustiqueError> {
     };
    
     for (index, mv) in rels.iter().enumerate() {
-        let version = &mv.mod_version.clone().unwrap_or(String::new()).to_string();
+        let version = &mv.mod_version.clone().unwrap().to_string();
         let game_versions = &mv.tags.clone().join(", ");
         let changelog = html2text::from_read(&mut mv.changelog.clone().unwrap_or(String::new()).as_bytes(), 100).map_err(|_| RustiqueError::SimpleError("html2txt failed".to_string()))?;